+++ src/widgets/notefilepathlabel.h
@@ -1,0 +1,29 @@
+/*
+ * Copyright (c) 2014-2024 Patrizio Bekerle -- <patrizio@bekerle.com>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; version 2 of the License.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
+ * for more details.
+ */
+
+#pragma once
+
+#include <QContextMenuEvent>
+#include <QLabel>
+#include <QString>
+#include <QWidget>
+
+class NoteFilePathLabel : public QLabel {
+    Q_OBJECT
+
+   public:
+    explicit NoteFilePathLabel(QWidget *parent = nullptr);
+
+   protected:
+    void contextMenuEvent(QContextMenuEvent *event) override;
+};